Moved WAP2 to right directory
